src/python/arc2face/requirements_controlnet.txt
absl-py==2.1.0
accelerate==0.29.3
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==1.3.1
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
chumpy==0.70
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
diffusers==0.23.0
easydict==1.13
exceptiongroup==1.2.1
face-alignment==1.4.1
facenet-pytorch==2.5.3
fastapi==0.110.2
ffmpy==0.3.2
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gradio==4.8.0
gradio_client==0.7.1
grpcio==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.17.3
humanfriendly==10.0
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
insightface==0.7.3
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.42.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numba==0.59.1
numpy==1.23.5
omegaconf==2.3.0
onnx==1.16.0
onnxruntime-gpu==1.17.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
orjson==3.10.1
packaging==24.0
pandas==2.2.2
peft==0.10.0
prettytable==3.10.0
protobuf==5.26.1
psutil==5.9.8
pydantic==2.7.1
pydantic_core==2.18.2
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytorch-lightning==1.4.9
pytz==2024.1
qudida==0.0.4
referencing==0.35.0
regex==2024.4.16
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.4.0
tifffile==2024.4.24
tokenizers==0.14.1
tomlkit==0.12.0
toolz==0.12.1
torchfile==0.1.0
torchmetrics==0.6.2
transformers==4.34.1
typer==0.12.3
tzdata==2024.1
uvicorn==0.29.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.2
yarl==1.9.4
zipp==3.18.1
